# openapi.yaml
openapi: 3.0.0
info:
  title: Complex API Demo
  version: 1.0.0
  description: A comprehensive API demonstration for Unkey deployment showcase
  contact:
    name: Unkey Team
    url: https://unkey.dev
    email: support@unkey.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: API server
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        data:
          oneOf:
            - type: object
            - type: array
            - type: 'null'
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
      required:
        - data
        - success
        - message
...
paths:
...
  /api/webhooks:
    get:
      summary: Get paginated list of webhooks
      tags:
        - Webhooks
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: Number of items per page
        - in: query
          name: active
          schema:
            type: boolean
          description: Filter by webhook status
      responses:
        '200':
          description: Paginated list of webhooks
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        url:
                          type: string
                          format: uri
                        events:
                          type: array
                          items:
                            type: string
                        active:
                          type: boolean
                        description:
                          type: string
                        lastTriggered:
                          type: string
                          format: date-time
                          nullable: true
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  success:
                    type: boolean
                  message:
                    type: string
    post:
      summary: Create a new webhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
                - events
              properties:
                url:
                  type: string
                  format: uri
                  example: https://example.com/webhook
                events:
                  type: array
                  items:
                    type: string
                  example:
                    - user.created
                    - order.completed
                secret:
                  type: string
                  description: Secret for webhook signature verification
                active:
                  type: boolean
                  default: true
                description:
                  type: string
      responses:
        '201':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request - invalid webhook data
  /api/webhooks/{id}:
    get:
      summary: Get webhook by ID
      tags:
        - Webhooks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook details
        '404':
          description: Webhook not found
    put:
      summary: Update webhook
      tags:
        - Webhooks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                active:
                  type: boolean
                description:
                  type: string
      responses:
        '200':
          description: Webhook updated successfully
        '404':
          description: Webhook not found
    delete:
      summary: Delete webhook
      tags:
        - Webhooks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Webhook deleted successfully
        '404':
          description: Webhook not found
  /api/webhooks/{id}/test:
    post:
      summary: Test webhook by sending a test payload
      tags:
        - Webhooks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Test webhook sent successfully
        '404':
          description: Webhook not found